requirements.txt
Flask==3.0.0
diskcache==5.6.1
line-bot-sdk==3.14.2
python-dotenv==1.0.0
requests>=2.32.3
gunicorn==21.2.0
biopython==1.81
googletrans==4.0.0-rc1